+++ pyp2spec/rpmversion.py
@@ -1,0 +1,63 @@
+# This script is a cut and modified version of pyreq2rpm.
+#
+# Copyright 2019 Gordon Messmer <gordon.messmer@gmail.com>
+#
+# Original: https://github.com/gordonmessmer/pyreq2rpm
+#
+# Permission is hereby granted, free of charge, to any person
+# obtaining a copy of this software and associated documentation files
+# (the "Software"), to deal in the Software without restriction,
+# including without limitation the rights to use, copy, modify, merge,
+# publish, distribute, sublicense, and/or sell copies of the Software,
+# and to permit persons to whom the Software is furnished to do so,
+# subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be
+# included in all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+# NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
+# BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+# ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+# CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+from packaging.version import parse
+
+class RpmVersion():
+    def __init__(self, version_id):
+        version = parse(version_id)
+        if isinstance(version._version, str):
+            self.version = version._version
+        else:
+            self.epoch = version._version.epoch
+            self.version = list(version._version.release)
+            self.pre = version._version.pre
+            self.dev = version._version.dev
+            self.post = version._version.post
+            # version.local is ignored as it is not expected to appear
+            # in public releases
+            # https://www.python.org/dev/peps/pep-0440/#local-version-identifiers
+
+    def is_legacy(self):
+        return isinstance(self.version, str)
+
+    def __str__(self):
+        if self.is_legacy():
+            return self.version
+        if self.epoch:
+            rpm_epoch = str(self.epoch) + ':'
+        else:
+            rpm_epoch = ''
+        rpm_version = '.'.join(str(x) for x in self.version)
+        if self.pre:
+            rpm_suffix = '~{}'.format(''.join(str(x) for x in self.pre))
+        elif self.dev:
+            rpm_suffix = '~~{}'.format(''.join(str(x) for x in self.dev))
+        elif self.post:
+            rpm_suffix = '^post{}'.format(self.post[1])
+        else:
+            rpm_suffix = ''
+        return '{}{}{}'.format(rpm_epoch, rpm_version, rpm_suffix)