+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 this was made the default option as of pyproject-rpm-macros 0-53
 - It's possible to declare archful package via CLI (`--archful, -a`) -
 the generated spec file will not contain `BuildArch: noarch` line
+- Package version compatible with PEP 440 is now converted to RPM scheme,
+version not following the specification remains unchanged.
+For that, a modified part of [pyreq2rpm](https://github.com/gordonmessmer/pyreq2rpm/blob/master/pyreq2rpm/pyreq2rpm.py) is used
 
 
 ## [0.3.3] - 2022-01-11